wm core

+++ src/status.c
@@ -13,7 +13,7 @@ wg_t status;
 void
 status_init(Window const win) {
   status = wg_init(win, 1, font.ch, 0);
-  static unsigned const STATUSMASK = 
+  static long const STATUSMASK = 
     PropertyChangeMask |
     ExposureMask;
   XSelectInput(dpy, status.win, STATUSMASK);
